chore: merge util and rustybeer libraries
this is to make handling conversions more easier for #109
