+++ Cargo.toml
@@ -2,6 +2,5 @@
 members = [
     "rustybeer",
     "rustybeer-cli",
-    "rustybeer-util",
     "rustybeer-server"
 ]
