+++ README.md
@@ -59,20 +59,15 @@ cargo test -- --nocapture <test name>
 - [**rustybeer**](rustybeer) - The folder containing everything for the business logic
   - [**src**](rustybeer/src) - The folder containing the business logic source code
     - [**calculators**](rustybeer/src/calculators) - The folder containing calculators to be used in lib or CLI tool
-  - [**tests**](rustybeer/tests) - The folder containing test source code
-    - [**calculators.rs**](rustybeer/tests/calculators.rs) The file containing the test source code for the calculators
   - [**Cargo.toml**](rustybeer/Cargo.toml) - The file containing build and dependency infomation
 - [**rustybeer-cli**](rustybeer-cli) - The folder containing everything for the CLI
   - [**src**](rustybeer-cli/src) - The folder containing the CLI source code
     - [**commands**](rustybeer-cli/src/commands) - The folder containing subcommands for CLI
     - [**main.rs**](rustybeer-cli/src/main.rs) - The file containing the main function
   - [**Cargo.toml**](rustybeer-cli/Cargo.toml)- The file containing build and dependency infomation
-- [**rustybeer-util**](rustybeer-util) - The folder containing extra utilities
-  - [**src**](rustybeer-util/src) - The folder containing the utilities source code
-  - [**Cargo.toml**](rustybeer-util/Cargo.toml) - The file containing build and dependency infomation
 - [**rustybeer-server**](rustybeer-server) - The folder containing the HTTP server implementation
   - [**src**](rustybeer-server/src) - The folder containing server source code
-    - [**handlers**](rustybeer-server/handlers) - HTTP request handlers
+    - [**handlers**](rustybeer-server/src/handlers) - HTTP request handlers
   - [**Cargo.toml**](rustybeer-server/Cargo.toml) - The file containing server build and dependency information
 - [**Cargo.toml**](Cargo.toml) - The file containing build and dependency infomation
 - [**CONTRIBUTING.md**](CONTRIBUTING.md) - Contribution guidelines for this repository
