+++ rustybeer-cli/Cargo.toml
@@ -6,8 +6,7 @@ edition = "2018"
 
 [dependencies]
 anyhow = "1.0"
-rustybeer = { version = "0.1.0", path = "../rustybeer"}
-rustybeer-util = { version = "0.1.0", path = "../rustybeer-util"}
+rustybeer = { path = "../rustybeer"}
 structopt = "0.3.20"
 chrono = "0.4"
