+++ rustybeer-server/Cargo.toml
@@ -6,7 +6,6 @@ edition = "2018"
 
 [dependencies]
 rustybeer = { path = "../rustybeer" }
-rustybeer-util = { version = "0.1.0", path = "../rustybeer-util"}
 rweb = { version = "0.13.0", features = ["openapi"] }
 serde = "1"
 tokio = "1"
