--- rustybeer-util/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "rustybeer-util"
-version = "0.1.0"
-authors = ["Heikki Hellgren <heiccih@gmail.com>", "mlatief", "Joseph Russell"]
-edition = "2018"
-
-[dependencies]
-approx = "0.5.0"
-measurements = "0.10.3"
-once_cell = "1.4.1"
-regex = "1.3.9"
-serde = { version = "1.0.116", features = ["derive"] }
-serde_json = "1.0.58"
