--- rustybeer-util/src/lib.rs
@@ -1,16 +1,0 @@
-//! Utilities for working with `rustybeer`:
-//! * List of beer styles that can be indexed and requested
-//! * List of hops and their AA%
-//! * List of yeasts and their properties
-//! * Units conversions from strings
-
-pub mod abv_calories;
-pub mod beer_styles;
-pub mod conversions;
-pub mod hops;
-pub mod yeasts;
-
-pub use measurements;
-
-mod macros;
-mod strings;
