+++ rustybeer/Cargo.toml
@@ -5,7 +5,9 @@ authors = ["Heikki Hellgren <heiccih@gmail.com>", "mlatief", "Joseph Russell"]
 edition = "2018"
 
 [dependencies]
-
-[dev-dependencies]
 approx = "0.5.0"
-rustybeer-util = { version = "0.1.0", path = "../rustybeer-util"}
+measurements = "0.10.3"
+once_cell = "1.4.1"
+regex = "1.3.9"
+serde = { version = "1.0.116", features = ["derive"] }
+serde_json = "1.0.58"
