+++ rustybeer/src/calculators/yeast_viability.rs
@@ -9,3 +9,18 @@ pub fn calculate_yv(days: f32) -> f32 {
 pub fn calculate_cc(cc: f32, days: f32) -> f32 {
     cc * (calculate_yv(days) / 100.0)
 }
+
+#[cfg(test)]
+pub mod tests {
+    use super::*;
+    use crate::assert_approx;
+    #[test]
+    fn yeast_viability() {
+        assert_approx!(970.0, calculate_cc(1000.0, 0.0));
+        assert_approx!(115.048_26, calculate_cc(123.45, 5.0));
+        assert_approx!(0.0, calculate_cc(9001.0, 3650.0));
+        assert_approx!(97.0, calculate_yv(0.0));
+        assert_approx!(65.004_616, calculate_yv(50.0));
+        assert_approx!(0.0, calculate_yv(3650.0));
+    }
+}
