+++ rustybeer/src/strings.rs
